Decimal precision

+++ Client/Records/ActiveOrderRecord.cs
@@ -9,10 +9,10 @@ namespace Client.Records
   public struct ActiveOrderRecord
   {
     public string Name { get; set; }
-    public double Size { get; set; }
-    public double Gain { get; set; }
-    public double OpenPrice { get; set; }
-    public double ClosePrice { get; set; }
+    public decimal Size { get; set; }
+    public decimal Gain { get; set; }
+    public decimal OpenPrice { get; set; }
+    public decimal ClosePrice { get; set; }
     public DateTime? Time { get; set; }
     public OrderSideEnum? Side { get; set; }
   }
